Add X2Y2 marketplace

+++ src/marketplaces/wyvern/lib/ArrayUtils.sol
@@ -2,25 +2,31 @@
 pragma solidity >=0.4.13;
 
 library ArrayUtils {
-
-    function guardedArrayReplace(bytes memory array, bytes memory desired, bytes memory mask)
-        internal
-        pure
-    {
+    function guardedArrayReplace(
+        bytes memory array,
+        bytes memory desired,
+        bytes memory mask
+    ) internal pure {
         require(array.length == desired.length);
         require(array.length == mask.length);
 
-        uint words = array.length / 0x20;
-        uint index = words * 0x20;
+        uint256 words = array.length / 0x20;
+        uint256 index = words * 0x20;
         assert(index / 0x20 == words);
-        uint i;
+        uint256 i;
 
         for (i = 0; i < words; i++) {
             /* Conceptually: array[i] = (!mask[i] && array[i]) || (mask[i] && desired[i]), bitwise in word chunks. */
             assembly {
                 let commonIndex := mul(0x20, add(1, i))
                 let maskValue := mload(add(mask, commonIndex))
-                mstore(add(array, commonIndex), or(and(not(maskValue), mload(add(array, commonIndex))), and(maskValue, mload(add(desired, commonIndex)))))
+                mstore(
+                    add(array, commonIndex),
+                    or(
+                        and(not(maskValue), mload(add(array, commonIndex))),
+                        and(maskValue, mload(add(desired, commonIndex)))
+                    )
+                )
             }
         }
 
@@ -31,13 +37,21 @@ library ArrayUtils {
             assembly {
                 let commonIndex := mul(0x20, add(1, i))
                 let maskValue := mload(add(mask, commonIndex))
-                mstore(add(array, commonIndex), or(and(not(maskValue), mload(add(array, commonIndex))), and(maskValue, mload(add(desired, commonIndex)))))
+                mstore(
+                    add(array, commonIndex),
+                    or(
+                        and(not(maskValue), mload(add(array, commonIndex))),
+                        and(maskValue, mload(add(desired, commonIndex)))
+                    )
+                )
             }
         } else {
             /* If the byte array is shorter than a word, we must unfortunately do the whole thing bytewise.
                (bounds checks could still probably be optimized away in assembly, but this is a rare case) */
             for (i = index; i < array.length; i++) {
-                array[i] = ((mask[i] ^ 0xff) & array[i]) | (mask[i] & desired[i]);
+                array[i] =
+                    ((mask[i] ^ 0xff) & array[i]) |
+                    (mask[i] & desired[i]);
             }
         }
     }
@@ -63,10 +77,10 @@ library ArrayUtils {
      * @param source uint to write
      * @return End memory index
      */
-    function unsafeWriteAddressWord(uint index, address source)
+    function unsafeWriteAddressWord(uint256 index, address source)
         internal
         pure
-        returns (uint)
+        returns (uint256)
     {
         assembly {
             mstore(index, source)
@@ -82,10 +96,10 @@ library ArrayUtils {
      * @param source uint to write
      * @return End memory index
      */
-    function unsafeWriteUint(uint index, uint source)
+    function unsafeWriteUint(uint256 index, uint256 source)
         internal
         pure
-        returns (uint)
+        returns (uint256)
     {
         assembly {
             mstore(index, source)
@@ -101,10 +115,10 @@ library ArrayUtils {
      * @param source uint8 to write
      * @return End memory index
      */
-    function unsafeWriteUint8(uint index, uint8 source)
+    function unsafeWriteUint8(uint256 index, uint8 source)
         internal
         pure
-        returns (uint)
+        returns (uint256)
     {
         assembly {
             mstore8(index, source)
@@ -120,10 +134,10 @@ library ArrayUtils {
      * @param source uint to write
      * @return End memory index
      */
-    function unsafeWriteUint8Word(uint index, uint8 source)
+    function unsafeWriteUint8Word(uint256 index, uint8 source)
         internal
         pure
-        returns (uint)
+        returns (uint256)
     {
         assembly {
             mstore(index, source)
@@ -139,10 +153,10 @@ library ArrayUtils {
      * @param source uint to write
      * @return End memory index
      */
-    function unsafeWriteBytes32(uint index, bytes32 source)
+    function unsafeWriteBytes32(uint256 index, bytes32 source)
         internal
         pure
-        returns (uint)
+        returns (uint256)
     {
         assembly {
             mstore(index, source)
